refactor: modularização e melhorias na estrutura do código
- Adiciona  para abstrair a lógica de salvamento no InfluxDB.
- Cria  para lidar com a lógica de obtenção e processamento dos níveis do rio.
- Move a função  para  e a expõe via init.
- Atualiza  para usar a nova estrutura de serviços e agendador ().
- Introduz  como arquivo inicial de configuração para o Grafana.
- Ajusta  para incluir o cache de pacotes Python ().
